Split finding boundary encode and decode responsibilities
