+++ apps/server/tests/shared/boundaries/test_finding_legacy_aliases.py
@@ -3,7 +3,8 @@
 from __future__ import annotations
 
 from vibesensor.domain import Finding, FindingEvidence, VibrationSource
-from vibesensor.shared.boundaries.finding import finding_from_payload, finding_payload_from_domain
+from vibesensor.shared.boundaries.finding_decoder import finding_from_payload
+from vibesensor.shared.boundaries.finding_encoder import finding_payload_from_domain
 
 
 class TestLegacySourceAlias:
